requirements.txt
streamlit==1.25.0
streamlit-drawable-canvas==0.9.2
opencv-python
cvzone
numpy
Pillow
